pyproject.toml
[project]
name = "minna-core"
version = "0.1.0"
description = "Local-first AI context engine. Your data stays on your machine."
authors = [
    {name = "Minna Team", email = "hello@minna.ai"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11,<3.14"
keywords = ["ai", "context", "local-first", "mcp", "vector-database", "slack", "github"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "fastembed>=0.7.4,<0.8.0",
    "mcp[cli]>=1.25.0,<2.0.0",
    "pydantic>=2.12.5,<3.0.0",
    "python-dotenv>=1.2.1,<2.0.0",
    "slack-sdk>=3.39.0,<4.0.0",
    "sqlite-vec>=0.1.6,<0.2.0",
    "requests>=2.31.0,<3.0.0",
    "keyring>=25.0.0,<26.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[project.scripts]
minna = "minna.cli:main"

[project.urls]
Homepage = "https://github.com/minna-ai/minna-core"
Documentation = "https://github.com/minna-ai/minna-core#readme"
Repository = "https://github.com/minna-ai/minna-core"
Issues = "https://github.com/minna-ai/minna-core/issues"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{include = "minna", from = "src"}]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
